requirements.txt
alf==0.7.1
APScheduler==3.6.1
Babel==2.3.4
boto3==1.7.72
django-all-access==0.9.0
Django==1.11.20
iso8601==0.1.11
mock==1.3.0
MySQL-python==1.2.5
netaddr==0.7.18
openpyxl==2.3.5
oslo.config==3.2.0
prettytable==0.7.2
pymongo==3.3.0
PyMySQL==0.8.0
python-dateutil==2.5.3
python-keystoneclient==2.1.2
python-swiftclient==2.5.0
pytz==2015.4
requests==2.19.1
SQLAlchemy==1.2.3
sqlparse==0.1.16
storm-keystone==0.1.3
wsgiref==0.1.2
